+++ package.json
@@ -50,7 +50,7 @@
     "@babel/preset-react": "^7.26.3",
     "@babel/preset-typescript": "7.9.0",
     "@babel/register": "7.25.9",
-    "@ryusenpai/shared-components": "1.14.4",
+    "@ryusenpai/shared-components": "1.14.5",
     "@testing-library/react": "16.2.0",
     "@types/express": "4.17.21",
     "@types/jest": "25.2.1",
