+++ src/components/header/index.module.css
@@ -1,16 +1,22 @@
 .navWrapper {
   display: flex;
+  justify-content: space-between;
   border-radius: 10px;
-  gap: 1.5rem;
   height: var(--header-height);
 
+  & > div:nth-of-type(1) {
+    display: flex;
+    align-items: center;
+    gap: 1.5rem;
+  }
+
   background-color: var(--color-primary);
-  padding: 1rem 2rem;
+  padding: 0.5rem 1.5rem;
   margin-bottom: var(--space-xs);
 
-  & svg {
-    width: 40px;
-    height: 40px;
+  & > div svg {
+    width: 50px;
+    height: 50px;
   }
 
   & button {
@@ -19,6 +25,16 @@
     color: #fff;
     padding-inline: var(--space-sm);
     border-radius: 15px;
+    height: 100%;
+  }
+}
+
+.settingsWrapper {
+  display: grid;
+  place-items: center;
+  & span svg {
+    width: 40px;
+    height: 40px;
   }
 }
 
@@ -90,6 +106,35 @@
   display: flex;
   align-items: center;
   gap: var(--space-xs);
+}
+
+.modalContent {
+  --dropdown-width: 130px;
+  & [role='listbox'] {
+    display: flex;
+    justify-content: space-between;
+    width: var(--dropdown-width);
+  }
+
+  ul {
+    width: var(--dropdown-width);
+  }
+
+  & li {
+    border: none;
+
+    &:last-child {
+      border-bottom: none;
+    }
+  }
+
+  & li[aria-selected='true'] {
+    color: #fff;
+  }
+
+  & svg {
+    fill: var(--color-primary);
+  }
 }
 
 @keyframes scrolling {
