+++ src/config/constants.ts
@@ -1,3 +1,9 @@
 import conversationConfig from './conversation.json'
 
 export const conversation = conversationConfig.conversation
+
+export const themeOptions = [
+  {text: 'default', value: 'default'},
+  {text: 'Sky Green', value: 'skyGreen'},
+  {text: 'Naruto', value: 'naruto'}
+]
